Add the modules picker
